requirements.txt
streamlit==1.28.2
openai==1.3.4
python-dotenv==1.0.0
